tests and clean up

+++ src/main/java/com/gleb/vinnikov/venue/subscriptions/services/impl/SubscriptionServiceImpl.java
@@ -2,10 +2,12 @@ package com.gleb.vinnikov.venue.subscriptions.services.impl;
 
 import com.gleb.vinnikov.venue.db.entities.User;
 import com.gleb.vinnikov.venue.db.entities.UserVenueSubscription;
+import com.gleb.vinnikov.venue.db.entities.Venue;
 import com.gleb.vinnikov.venue.db.repos.UserVenueSubscriptionsRepo;
 import com.gleb.vinnikov.venue.db.repos.VenueRepo;
 import com.gleb.vinnikov.venue.subscriptions.api.UserResponse;
 import com.gleb.vinnikov.venue.subscriptions.services.SubscriptionService;
+import com.gleb.vinnikov.venue.users.api.UserResponsePublic;
 import com.gleb.vinnikov.venue.utils.GeneralUtils;
 import com.gleb.vinnikov.venue.utils.Message;
 import com.gleb.vinnikov.venue.venues.api.VenueResponse;
@@ -17,6 +19,7 @@ import org.springframework.data.domain.Example;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -57,21 +60,30 @@ public class SubscriptionServiceImpl implements SubscriptionService {
     }
 
     @Override
-    public List<UserResponse> getVenueSubscribersByVenueId(@NonNull UUID venueId) {
+    public List<UserResponsePublic> getVenueSubscribersByVenueId(@NonNull UUID venueId) {
+        if (!venueRepo.existsById(venueId)) {
+            throw new IllegalArgumentException("No venue with give id");
+        }
         return GeneralUtils.mapList(
                 userVenueSubscriptionsRepo.findUsersByVenueId(venueId),
-                this::subscriptionToUserResponse);
+                this::subscriptionToUserResponsePublic);
     }
 
     @Override
-    public List<UserResponse> getVenueSubscribersByVenueIdName(@NonNull String venueIdName) {
+    public List<UserResponsePublic> getVenueSubscribersByVenueIdName(@NonNull String venueIdName) {
+        if (!venueRepo.existsByIdName(venueIdName)) {
+            throw new IllegalArgumentException("No venue with give idName");
+        }
         return GeneralUtils.mapList(
                 userVenueSubscriptionsRepo.findUsersByVenueIdName(venueIdName),
-                this::subscriptionToUserResponse);
+                this::subscriptionToUserResponsePublic);
     }
 
     @Override
     public Message unsubscribeFromVenue(@NonNull UUID userId, @NonNull UUID venueId) {
+        if (!venueRepo.existsById(venueId)) {
+            throw new IllegalArgumentException("No venue with give id");
+        }
         return userVenueSubscriptionsRepo.findIdByUserIdAndVenueId(userId, venueId)
                 .map(id -> {
                     userVenueSubscriptionsRepo.deleteById(id);
@@ -87,8 +99,8 @@ public class SubscriptionServiceImpl implements SubscriptionService {
                 .orElseThrow(() -> new IllegalArgumentException("No venue with given idName."));
     }
 
-    private UserResponse subscriptionToUserResponse(UserVenueSubscription.SubscriptionUserOnly u) {
-        return new UserResponse(u.getUser().getId(), u.getUser().getUsername());
+    private UserResponsePublic subscriptionToUserResponsePublic(UserVenueSubscription u) {
+        return new UserResponsePublic(u.getUser().getId(), u.getUser().getUsername());
     }
 
 }
